+++ tests/testthat/test_peak-list-functions.R
@@ -1,4 +1,4 @@
-test_that("combinePeaks works", {
+test_that("combinePeaksData works", {
     set.seed(123)
     mzs <- seq(1, 20, 0.1)
     ints1 <- abs(rnorm(length(mzs), 10))
@@ -16,15 +16,20 @@ test_that("combinePeaks works", {
                 intensity = ints3)
 
     ## With tolerance = 0 and ppm = 0 we expect to get just a joined matrix.
-    res <- combinePeaks(list(p1, p2, p3))
+    expect_warning(combinePeaks(list(p1, p2, p3)), "deprecated")
+    res <- combinePeaksData(list(p1, p2, p3))
     p <- rbind(p1, p2, p3)
     p <- p[order(p[, 1]), ]
     expect_identical(p, res)
 
     ## With tolerance = 0.05 we group all "triplets"
-    res <- combinePeaks(list(p1, p2, p3), tolerance = 0.05, mzFun = min,
+    res <- combinePeaksData(list(p1, p2, p3), tolerance = 0.05, mzFun = min,
                         intensityFun = min)
     expect_true(nrow(res) == length(mzs))
+    expect_warning(res2 <- combinePeaks(list(p1, p2, p3), tolerance = 0.05,
+                                        mzFun = min, intensityFun = min),
+                   "deprecated")
+    expect_equal(res, res2)
 
     p <- do.call(rbind, lapply(1:nrow(res), function(i) {
         c(mz = min(p1[i, 1], p2[i, 1], p3[i, 1]),
@@ -38,7 +43,7 @@ test_that("combinePeaks works", {
                 intensity = c(12, 22, 32, 42))
     p3 <- cbind(mz = c(23 - ppm(23, 5), 24), intensity = c(13, 23))
 
-    res <- combinePeaks(list(p1, p2, p3), ppm = 10)
+    res <- combinePeaksData(list(p1, p2, p3), ppm = 10)
     expect_true(nrow(res) == 6)
     expect_true(res[1, 2] == 12)
     expect_true(res[2, 2] == mean(c(10, 22, 13)))
@@ -54,7 +59,7 @@ test_that("combinePeaks works", {
     expect_true(res[5, 1] == mean(p1[3, 1]))
     expect_true(res[6, 1] == mean(c(p1[4, 1], p2[4, 1])))
 
-    res <- combinePeaks(list(p1, p2, p3), ppm = 10, intensityFun = median,
+    res <- combinePeaksData(list(p1, p2, p3), ppm = 10, intensityFun = median,
                         mzFun = max)
     expect_true(nrow(res) == 6)
     expect_true(res[1, 2] == 12)
@@ -76,15 +81,16 @@ test_that("combinePeaks works", {
     p2 <- cbind(mz = c(17, 45.1, 63.9, 70.2), intensity = c(11, 21, 31, 41))
     p3 <- cbind(mz = c(12.1, 44.9, 63), intensity = c(12, 22, 32))
 
-    expect_identical(combinePeaks(list(p1), peaks = "intersect"), p1)
+    expect_identical(combinePeaksData(list(p1), peaks = "intersect"), p1)
 
-    res <- combinePeaks(list(p1, p2, p3), peaks = "intersect", minProp = 0.2)
+    res <- combinePeaksData(list(p1, p2, p3), peaks = "intersect",
+                            minProp = 0.2)
     expect_equal(res[, 1], sort(c(p1[, 1], p2[, 1], p3[, 1])))
     idx <- order(c(p1[, 1], p2[, 1], p3[, 1]))
     expect_equal(res[, 2], c(p1[, 2], p2[, 2], p3[, 2])[idx])
 
-    res_2 <- combinePeaks(list(p1, p2, p3), peaks = "union", main = 3,
-                          tolerance = 1)
+    res_2 <- combinePeaksData(list(p1, p2, p3), peaks = "union", main = 3,
+                              tolerance = 1)
     expect_true(nrow(res_2) == 3)
     expect_equal(unname(res_2[1, 1]), mean(c(12, 12.1)))
     expect_equal(unname(res_2[2, 1]), mean(c(45, 45.1, 44.9)))
@@ -93,23 +99,24 @@ test_that("combinePeaks works", {
     expect_equal(unname(res_2[2, 2]), mean(c(20, 21, 22)))
     expect_equal(unname(res_2[3, 2]), mean(c(30, 31, 32)))
 
-    res <- combinePeaks(list(p1, p2, p3), peaks = "intersect")
+    res <- combinePeaksData(list(p1, p2, p3), peaks = "intersect")
     expect_true(nrow(res) == 0)
 
-    res <- combinePeaks(list(p1, p2, p3), tolerance = 0.1, peaks = "intersect")
+    res <- combinePeaksData(list(p1, p2, p3), tolerance = 0.1,
+                            peaks = "intersect")
     expect_equal(res[, 1], c(mean(c(12, 12.1)), mean(c(45, 45.1, 44.9)),
                              mean(c(64, 63.9))))
     expect_equal(res[, 2], c(mean(c(10, 12)), mean(c(20, 21, 22)),
                              mean(c(30, 31))))
 
-    res <- combinePeaks(list(p1, p2, p3), tolerance = 0.1, mzFun = min,
-                        intensityFun = max, peaks = "intersect")
+    res <- combinePeaksData(list(p1, p2, p3), tolerance = 0.1, mzFun = min,
+                            intensityFun = max, peaks = "intersect")
     expect_equal(res[, 1], c(min(c(12, 12.1)), min(c(45, 45.1, 44.9)),
                              min(c(64, 63.9))))
     expect_equal(res[, 2], c(max(c(10, 12)), max(c(20, 21, 22)),
                              max(c(30, 31))))
 
-    expect_error(combinePeaks(list(p1, p2, p3), main = 5), "has to be")
+    expect_error(combinePeaksData(list(p1, p2, p3), main = 5), "has to be")
 })
 
 test_that(".peaks_compare works", {